Add strong parser tests.
Most don't pass (and are disabled), but should.  Unfortunately, adding a word boundary (\b) to the strong regex works for these tests, but somehow the full parser then breaks.  The regex that I believed should work is added as a comment, for anyone wishing to make things work going forward.